Make Document.Info members mutable
I'm working on a project that requires lots of individual mutation of the entire Document structure, and I couldn't find a particular reason why all the Document.Info fields are immutable. The rest of the structure seems to be mutable and I couldn't find a particular reason why these members need to be immutable. Wondering if there's something I'm overlooking. Thanks so much for this awesome package 😀